Major_Project/requirements.txt
streamlit
python-dotenv
google-generativeai
SpeechRecognition
gTTS
pydub
pyaudio
google-auth
speechrecognition
gtts
playsound

